Destroy Delay
Implemented a new system for determining when to destroy a spawned effect game object.   Now it is based off of seconds(real-time) and not frames which was creating a big difference in terms of gameplay depending on whether you were playing on a PC or Mac.   Aka I discovered that you shouldn't base ANYTHING off of frame count because everyone's experience will differ. <3

Also I'm proud of you for setting aside some time to work on this.  It feels like we're growing and getting better even if it's little baby steps.